+++ Sources/Engine/Models/Model.cpp
@@ -481,7 +481,7 @@ ModelTextureVertex::ModelTextureVertex(void)
 //------------------------------------------ WRITE
 void ModelPolygonVertex::Write_t( CTStream *pFile)  // throw char *
 {
-  STUBBED("64-bit issue");
+  STUBBED("64-bit issue"); // DG: probably ok, because PtrToIndices() should have been called before this
   (*pFile) << (INDEX) (size_t) mpv_ptvTransformedVertex;
   (*pFile) << (INDEX) (size_t) mpv_ptvTextureVertex;
 }
@@ -490,7 +490,7 @@ void ModelPolygonVertex::Read_t( CTStream *pFile) // throw char *
 {
   INDEX itmp;
 
-  STUBBED("64-bit issue");
+  STUBBED("64-bit issue"); // DG: probably ok, because IndicesToPtrs() should be called afterwards
   (*pFile) >> itmp;
   mpv_ptvTransformedVertex = (struct TransformedVertexData *) (size_t) itmp;
   (*pFile) >> itmp;
@@ -1040,14 +1040,14 @@ void CModelData::PtrsToIndices()
           if( it2.Current().mpv_ptvTransformedVertex == &md_TransformedVertices[ j])
             break;
         }
-        it2.Current().mpv_ptvTransformedVertex = (struct TransformedVertexData *) j;
+        it2.Current().mpv_ptvTransformedVertex = (struct TransformedVertexData *)(size_t)j;
 
         for( j=0; j<md_MipInfos[ i].mmpi_TextureVertices.Count(); j++)
         {
           if( it2.Current().mpv_ptvTextureVertex == &md_MipInfos[ i].mmpi_TextureVertices[ j])
             break;
         }
-        it2.Current().mpv_ptvTextureVertex = (ModelTextureVertex *) j;
+        it2.Current().mpv_ptvTextureVertex = (ModelTextureVertex *)(size_t)j;
       }
     }
   }
@@ -1068,10 +1068,10 @@ void CModelData::IndicesToPtrs()
       FOREACHINSTATICARRAY(it1.Current().mp_PolygonVertices, ModelPolygonVertex, it2)
       {
         struct ModelPolygonVertex * pMPV = &it2.Current();
-        STUBBED("64-bit issue");
+        STUBBED("64-bit issue"); // DG: probably ok, the pointers really contain indices from PtrToIndices()
         j = (INDEX) (size_t) it2.Current().mpv_ptvTransformedVertex;
         it2.Current().mpv_ptvTransformedVertex = &md_TransformedVertices[ j];
-        STUBBED("64-bit issue");
+        STUBBED("64-bit issue"); // DG: probably ok, the pointers really contain indices from PtrToIndices()
         j = (INDEX) (size_t) it2.Current().mpv_ptvTextureVertex;
         it2.Current().mpv_ptvTextureVertex = &md_MipInfos[ i].mmpi_TextureVertices[ j];
       }
